Me logic+wallets (#29)
* added db methods to the engine to add and remove funds from wallets

* forgot to run the build first

* updated schemas for buy/sell orders

* finished logic for market buys and sell order handling. still needs testing and error handling

* fixed broken import

* Added function to take money from buyer and give to sellers in one database transaction

* updated schemas for engine stock_id

* basic buy/sell working

* partial buys working

* Print exception

* fix indentation

* import

* fix import statement

* fixed bug with adding money to seller wallet

* added error for when sell orders cant fill buy order. also changed heap state handling when transaction fails

* refactoring + removing print statements

* changed error handling from print statement

---------

+++ schemas/transaction.py
@@ -3,7 +3,7 @@ from datetime import datetime
 
 
 class AddMoneyRequest(BaseModel):
-    amount: float
+    amount: int
 
 
 class WalletTxResult(BaseModel):
@@ -18,3 +18,4 @@ class PortfolioResult(BaseModel):
     stock_id: int
     stock_name: str
     quantity_owned: int
+
